Create myarea.py
myarea has a couple of things it can do to SVG paths. 
first of all it takes only normalized path elements, that means they should be made of cubic berzier curves and lines. ('c', 'l', 'C' or/and 'L')
and you have to apply regular expressions and pass data as 'M int int c int int int int int int', so that means, you have to pass whatever goes inside 'd' inside your path element. 

'curvy' converts relative to absolute coordinates as well as it converts a line into a cubic brezier curve, which is still a line. 
this is useful if you want to calculate the area. it returns a list of lists, where each list is a cubic brezier curve. in the format 
[int int C int int int int int int]
 
'areacal' takes the list of lists made above and calculates the area under each brezier curve, which also gives the area enclosed inside a curve made of many cubic brezier curves.